Add qrc file to build if not debug version
If the configuration is set to Debug, this will add the qrc file to the
build and set the DEBUG_BUILD variable to tell the main application not
to add the qrc file as an import path.  Otherwise, the qrc file is added
to the build and the qrc file is added to the import path.