Only perform install with Release configuration
Because the Debug version of the executable does not contain the
required qml resources, it makes no sense to install that version of the
executable.  This change doesn't prevent someone from attempting to
install the debug version, but it will not actually install anything
unless the Release version is specified.