Add portable support for icons
Linux icon files specifications are described in
https://specifications.freedesktop.org/icon-theme-spec/latest/
That specification states that icon files should be installed in standard
locations and have specified file formats.  This uses the existing data
structure and adds several sizes of icons in png format based on the svg
version and automatically installs all of them in the appropriate
locations.  This fixes issue #36.

Signed-off-by: Ed Beroset <beroset@ieee.org>